requirements.txt
openai
python-dotenv
google-generativeai
langchain-openai
langchain-google-genai
langchain-ollama
langsmith
streamlit
langchain-aws
scikit-learn